+++ src/api/index.ts
@@ -3,3 +3,5 @@ export { DomainsApi } from './DomainsApi';
 export { LogsApi } from './LogsApi';
 export { MessagesApi } from './MessagesApi';
 export { SuppressionsApi } from './SuppressionsApi';
+export { EmailValidationApi } from './EmailValidationApi';
+export { WebhooksApi } from './WebhooksApi';