+++ src/models/index.ts
@@ -3,9 +3,13 @@ export { Domain } from './Domain';
 export { LogEntry } from './LogEntry';
 export { Pagination } from './Pagination';
 export { Suppression } from './Suppression';
+export { EmailValidationResult } from './EmailValidationResult';
+export { Webhook } from './Webhook';
 
 export type { CredentialData } from './Credential';
 export type { DomainData } from './Domain';
 export type { LogEntryData } from './LogEntry';
 export type { PaginationData } from './Pagination';
 export type { SuppressionData } from './Suppression';
+export type { EmailValidationResultData } from './EmailValidationResult';
+export type { WebhookData } from './Webhook';